Django_CRUD/django_CRUD/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
altgraph==0.17.3
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
coreapi==2.3.3
coreschema==0.0.4
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.4
django-filter==22.1
django-ratelimit==4.0.0
django-seed==0.3.1
django-simple-tags==0.6.0
django-user-agents==0.4.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
drf-yasg==1.21.4
executing==1.2.0
Faker==15.3.4
fastjsonschema==2.16.2
Flask==2.2.3
fqdn==1.5.1
frozenlist==1.3.3
future==0.18.2
geoip2==4.6.0
gunicorn==20.1.0
h11==0.14.0
idna==3.4
inflection==0.5.1
ipykernel==6.21.1
ipython==8.9.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.0
jupyter-events==0.6.3
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyter_server==2.3.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
maxminddb==2.2.0
mistune==2.0.5
multidict==6.0.4
mysqlclient==2.1.1
nbclassic==0.5.2
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
outcome==1.2.0
packaging==23.0
pandocfilters==1.5.0
parso==0.8.3
pefile==2022.5.30
pickleshare==0.7.5
platformdirs==2.6.2
prometheus-client==0.16.0
prompt-toolkit==3.0.36
psutil==5.9.4
psycopg2==2.9.5
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.14.0
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.6
pytz==2022.7
pywin32==305
pywin32-ctypes==0.2.0
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
redis==4.5.1
regex==2022.10.31
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
ruamel.yaml==0.17.21
selenium==4.7.2
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sqlparse==0.4.3
stack-data==0.6.2
terminado==0.17.1
text-unidecode==1.3
tinycss2==1.2.1
toml==0.10.2
toposort==1.7
tornado==6.2
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.9.2
typing-extensions==3.10.0.0
tzdata==2022.7
ua-parser==0.16.1
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.13
user-agents==2.2.0
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.5
wsproto==1.2.0
yarl==1.8.2
